backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
SQLAlchemy
python-dotenv
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator
python-multipart
